+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- **[BREAKING]** `getElementSubmissionValue` `form` parameter to `formElements`
+
 ## [0.2.5] - 2021-11-19
 
 ### Fixed